tide-up...

+++ Test_examples_ZVCV.py
@@ -4,11 +4,12 @@ import numpy as np
 import sys
 import os
 import subprocess
+from os.path import exists
 
 
 # Parameters
 # output_path = 'performance-tests-cmdstan/ZVCV_comparison/'
-if 0:
+if 1:
     example_list_path = 'list_of_examples_2.txt'
     output_path = 'ZVCV-results-2/'
     terminal_path = 'ZVCV-messages-2/terminal-messages/'
@@ -37,12 +38,12 @@ for i, example_path in enumerate(list_examples):
 
     file_dir = example_path.replace('performance-tests-cmdstan/', '')
     output_filename = file_dir.replace('/', '@')
-    terminal_m_output = open(terminal_path + output_filename + '.txt', 'w')
-    error_m_output = open(error_path + output_filename + '.txt', 'w')
 
-    subprocess.run(['python', 'PerformanceTest_ZVCV.py',
-                   file_dir, output_path], stdout=terminal_m_output, stderr=error_m_output, text=True)
-    terminal_m_output.close()
-    error_m_output.close()
+    if exists(output_path + output_filename + '.csv') is False:
+        terminal_m_output = open(terminal_path + output_filename + '.txt', 'w')
+        error_m_output = open(error_path + output_filename + '.txt', 'w')
+        subprocess.run(['python', 'PerformanceTest_ZVCV.py', file_dir, output_path], stdout=terminal_m_output, stderr=error_m_output, text=True)
+        terminal_m_output.close()
+        error_m_output.close()
